build image with gitlab-ci

+++ build-es.sh
@@ -16,17 +16,22 @@ rm -rf $PWD/docker/elasticsearch/data || true
 
 # build and launch custom ES image with local mount
 docker build ./docker/elasticsearch -t cdtn-es
-docker run -d -v $TMP_DATA_PATH:/usr/share/elasticsearch/data -p $TMP_ES_LOCAL_PORT:9200 --name cdtn-es cdtn-es
+docker run -d -v $TMP_DATA_PATH:/usr/share/elasticsearch/data -p $TMP_ES_LOCAL_PORT:9200 -e ES_JAVA_OPTS="-Xms2048m -Xmx2048m" -e TAKE_FILE_OWNERSHIP=1 --name cdtn-es cdtn-es
 
 # build python ES ingestion script
 docker build ./packages/code-du-travail-data -t cdtn-data
 
 # wait ES
-retry=60
+retry=3
 while
   ! curl -sS "http://127.0.0.1:$TMP_ES_LOCAL_PORT/_cat/health?h=status" &&
   [[ $(( retry-- )) -gt 0 ]];
 do echo "Waiting for Elasticsearch to go Green ($retry)" ; sleep 1 ; done ;
+if [ $retry -eq 0 ]
+then
+  echo "ERROR: cant reach ES at http://127.0.0.1:$TMP_ES_LOCAL_PORT"
+  exit 1
+fi
 
 # launch python ES ingestion script
 docker run --rm --network=host -e ELASTICSEARCH_URL="http://127.0.0.1:$TMP_ES_LOCAL_PORT" cdtn-data
@@ -35,7 +40,5 @@ docker run --rm --network=host -e ELASTICSEARCH_URL="http://127.0.0.1:$TMP_ES_LO
 docker stop cdtn-es
 
 # build a new ES docker image with produced data
-docker build ./docker/elasticsearch -f ./docker/elasticsearch/Dockerfile-with-data -t cdtn-es-2
-
-
-
+docker build ./docker/elasticsearch -f ./docker/elasticsearch/Dockerfile-with-data -t $IMAGE_NAME:$CI_COMMIT_SHA $CONTEXT
+docker push $IMAGE_NAME
